style(server): fix lint issues across 36 files
- Remove unnecessary protocol.UUID() conversions (type alias for string)
- Remove unnecessary string() conversions on protocol.UUID fields
- Fix goimports grouping (session_memory_extractor.go, server.go, servicecontext.go)
- Simplify slice literal ([]T{T{...}} to []T{{...}})
- Remove unnecessary uint32() conversion in update publisher
- Update misspell locale to accept both British and American English
  (project consistently uses British English: cancelled, centralised)

+++ internal/updates/update_publisher.go
@@ -283,7 +283,7 @@ func (u *UpdatePublisher) publishUpdates(ctx context.Context, items []UpdatePubl
 			}
 			logger.Debug(ctx, "[UpdatePublisher] published update",
 				zap.String("channel", item.Channel),
-				zap.Uint32("offset", uint32(update.Offset)),
+				zap.Uint32("offset", update.Offset),
 				zap.Int("items", len(update.Items)))
 		}
 	}
